Split up tests into something more maintainable
