+++ deno.json
@@ -22,6 +22,8 @@
     "test": "deno task fmt-check && deno task lint && deno task unit",
     "cov": "deno task test && deno coverage cov_profile",
     "lcov": "deno coverage --lcov --output=cov.lcov cov_profile",
-    "report": "deno task lcov && genhtml -o html_cov cov.lcov && open html_cov/index.html"
+    "cov-check": "deno task cov && deno task lcov && lcov --summary cov.lcov --fail-under-lines 100",
+    "report": "deno task lcov && genhtml -o html_cov cov.lcov && open html_cov/index.html",
+    "pre-commit": "deno task cov-check"
   }
 }
