+++ tests/instructions/instructions.test.ts
@@ -1,30 +1,7 @@
 import { Definition } from '../../definitions/definitions.ts';
 import { ts, tsm } from '../../deps.ts';
 import { generateInstructions } from '../../instructions/instructions.ts';
-import { InstructionType } from '../../instructions/instructions.type.ts';
-import { assertIsError, assertSnapshot, assertThrows, blocks } from '../../test.deps.ts';
-import { sanitiseInstructions, TestDefinition } from './instructions-tests.ts';
-
-function createTestName(
-  output: string,
-  ...[initialCondition, ...conditions]: string[]
-): string {
-  let testName = output;
-  if (initialCondition) {
-    testName = [testName, initialCondition].join(' ');
-  }
-
-  if (conditions.length > 0) {
-    const middleConditions = conditions.slice(0, conditions.length - 1);
-    const finalCondition = conditions[conditions.length - 1];
-    if (middleConditions.length > 0) {
-      testName = [testName, ...middleConditions].join(', ');
-    }
-    testName = [testName, 'and', finalCondition].join(' ');
-  }
-
-  return testName;
-}
+import { assertIsError, assertThrows, blocks } from '../../test.deps.ts';
 
 blocks.describe('Instructions', () => {
   blocks.describe('generateInstructions', () => {
@@ -81,497 +58,5 @@ blocks.describe('Instructions', () => {
         );
       },
     );
-  });
-
-  blocks.describe(`ADD Instruction`, () => {
-    const definitions: TestDefinition[] = [
-      {
-        name: createTestName(
-          'should define an ADD instruction if',
-          'the field is an array of nodes',
-          'the field nodes don\'t exist',
-          'no instructions are specified on the field definition',
-        ),
-        input: {
-          kind: ts.SyntaxKind.SourceFile,
-          statements: [
-            {
-              kind: ts.SyntaxKind.FunctionDeclaration,
-              parameters: [],
-            },
-          ],
-        },
-      },
-      {
-        name: createTestName(
-          'should define an ADD instruction if',
-          'the field is an array of nodes',
-          'the field nodes don\'t exist',
-          'instructions are specified on the field definition',
-        ),
-        input: {
-          kind: ts.SyntaxKind.SourceFile,
-          statements: [
-            {
-              __instructions: {
-                id: 'some-instruction',
-                rules: [],
-              },
-              kind: ts.SyntaxKind.FunctionDeclaration,
-              parameters: [],
-            },
-          ],
-        },
-      },
-      {
-        name: createTestName(
-          'should define an ADD instruction if',
-          'the field is an array of nodes',
-          'the field nodes exist',
-          'no instructions are specified on the field definition',
-        ),
-        input: {
-          kind: ts.SyntaxKind.SourceFile,
-          statements: [
-            {
-              kind: ts.SyntaxKind.FunctionDeclaration,
-              parameters: [],
-            },
-          ],
-        },
-        sourceFileContents: `
-          export function hello() {}
-        `,
-      },
-      {
-        name: createTestName(
-          'should define an ADD instruction if',
-          'the field is an array of nodes',
-          'the field node is not found using the defined instruction ID',
-        ),
-        input: {
-          kind: ts.SyntaxKind.SourceFile,
-          statements: [
-            {
-              __instructions: {
-                // Looks like there is an AST
-                id: 'name.escapedText="hello"',
-              },
-              kind: ts.SyntaxKind.FunctionDeclaration,
-              name: {
-                kind: ts.SyntaxKind.Identifier,
-                // TODO: can we use bindings here?
-                text: 'hello',
-              },
-              parameters: [],
-            },
-          ],
-        },
-        sourceFileContents: `
-          export function foo() {}
-        `,
-      },
-      {
-        name: createTestName(
-          'should not define an ADD instruction if',
-          'the field is an array of nodes',
-          'the field node is found using the defined instruction ID',
-        ),
-        input: {
-          kind: ts.SyntaxKind.SourceFile,
-          statements: [
-            {
-              __instructions: {
-                // TODO: Looks like there is an AST we can use instead
-                id: 'name.text="hello"',
-              },
-              kind: ts.SyntaxKind.FunctionDeclaration,
-              name: {
-                kind: ts.SyntaxKind.Identifier,
-                // TODO: can we use bindings here when we compile a template?
-                text: 'hello',
-              },
-              parameters: [],
-            },
-          ],
-        },
-        sourceFileContents: `
-          export function hello() {}
-        `,
-      },
-      {
-        name: createTestName(
-          'should define an ADD instruction if',
-          'the field is an array of nodes',
-          'the field is a nested definition',
-        ),
-        input: {
-          kind: ts.SyntaxKind.SourceFile,
-          statements: [
-            {
-              __instructions: {
-                id: 'name.text="hello"',
-              },
-              kind: ts.SyntaxKind.FunctionDeclaration,
-              name: {
-                kind: ts.SyntaxKind.Identifier,
-                // TODO: can we use bindings here when we compile a template?
-                text: 'hello',
-              },
-              modifiers: [
-                { kind: ts.SyntaxKind.ExportKeyword },
-                { kind: ts.SyntaxKind.AsyncKeyword },
-              ],
-              parameters: [],
-            },
-          ],
-        },
-        sourceFileContents: `
-          function hello() {}
-        `,
-      },
-    ];
-
-    for (const definition of definitions) {
-      blocks.it(definition.name, async (t) => {
-        // Arrange
-        const project = new tsm.Project();
-        const sourceFile = project.createSourceFile(
-          `${crypto.randomUUID()}.ts`,
-          definition.sourceFileContents,
-        );
-
-        // Act
-        const result = generateInstructions(sourceFile, definition.input);
-
-        // Assert
-        await assertSnapshot(t, sanitiseInstructions(result));
-      });
-    }
-  });
-
-  blocks.describe(`SET Instruction`, () => {
-    const definitions: TestDefinition[] = [
-      {
-        name: createTestName(
-          'should define an SET instruction if',
-          'the field is a node',
-          'the field node doesn\'t exist',
-          'no instructions are specified on the field definition',
-        ),
-        input: {
-          kind: ts.SyntaxKind.SourceFile,
-          statements: [
-            {
-              __instructions: {
-                id: 'name.text="hello"',
-              },
-              kind: ts.SyntaxKind.FunctionDeclaration,
-              parameters: [],
-              type: {
-                kind: ts.SyntaxKind.VoidKeyword,
-              },
-            },
-          ],
-        },
-        sourceFileContents: `
-          export function hello() {}
-        `,
-      },
-      {
-        name: createTestName(
-          'should define an SET instruction if',
-          'the field is a node',
-          'the field node is not found using the defined instruction ID',
-        ),
-        input: {
-          kind: ts.SyntaxKind.SourceFile,
-          statements: [
-            {
-              __instructions: {
-                id: 'name.text="hello"',
-              },
-              kind: ts.SyntaxKind.FunctionDeclaration,
-              parameters: [],
-              type: {
-                __instructions: {
-                  id: 'kind=114',
-                },
-                kind: ts.SyntaxKind.VoidKeyword,
-              },
-            },
-          ],
-        },
-        sourceFileContents: `
-          export function hello(): string {}
-        `,
-      },
-      {
-        name: createTestName(
-          'should not define an SET instruction if',
-          'the field is a node',
-          'the field node exists',
-          'no instructions are specified on the field definition',
-        ),
-        input: {
-          kind: ts.SyntaxKind.SourceFile,
-          statements: [
-            {
-              __instructions: {
-                id: 'name.text="hello"',
-              },
-              kind: ts.SyntaxKind.FunctionDeclaration,
-              parameters: [],
-              type: {
-                kind: ts.SyntaxKind.VoidKeyword,
-              },
-            },
-          ],
-        },
-        sourceFileContents: `
-          export function hello(): void {}
-        `,
-      },
-      {
-        name: createTestName(
-          'should not define an SET instruction if',
-          'the field is a node',
-          'the field node is found using the defined instruction ID',
-        ),
-        input: {
-          kind: ts.SyntaxKind.SourceFile,
-          statements: [
-            {
-              __instructions: {
-                id: 'name.text="hello"',
-              },
-              kind: ts.SyntaxKind.FunctionDeclaration,
-              parameters: [],
-              type: {
-                __instructions: {
-                  id: 'kind=114',
-                },
-                kind: ts.SyntaxKind.VoidKeyword,
-              },
-            },
-          ],
-        },
-        sourceFileContents: `
-          export function hello(): void {}
-        `,
-      },
-    ];
-
-    for (const definition of definitions) {
-      blocks.it(definition.name, async (t) => {
-        // Arrange
-        const project = new tsm.Project();
-        const sourceFile = project.createSourceFile(
-          `${crypto.randomUUID()}.ts`,
-          definition.sourceFileContents,
-        );
-
-        // Act
-        const result = generateInstructions(sourceFile, definition.input);
-
-        // Assert
-        await assertSnapshot(t, sanitiseInstructions(result));
-      });
-    }
-  });
-
-  blocks.describe(`INSERT Instruction`, () => {
-    const definitions: TestDefinition[] = [
-      {
-        name: createTestName(
-          'should define an INSERT instruction if',
-          'the field is an array of nodes',
-          'a rule evaluates to an INSERT instruction',
-          'a numerical index is defined for the INSERT position',
-        ),
-        input: {
-          kind: ts.SyntaxKind.SourceFile,
-          statements: [
-            {
-              __instructions: {
-                rules: [{
-                  instruction: InstructionType.INSERT,
-                  condition: 'name.text="hello"',
-                  index: 0,
-                }],
-              },
-              kind: ts.SyntaxKind.FunctionDeclaration,
-              name: {
-                kind: ts.SyntaxKind.Identifier,
-                text: 'hello',
-              },
-              parameters: [],
-            },
-          ],
-        },
-        sourceFileContents: `
-          function hello() {}
-        `,
-      },
-      {
-        name: createTestName(
-          'should define an INSERT instruction if',
-          'the field is an array of nodes',
-          'a rule evaluates to an INSERT instruction',
-          'a string index is evaluated to an integer for the INSERT position',
-        ),
-        input: {
-          kind: ts.SyntaxKind.SourceFile,
-          statements: [
-            {
-              __instructions: {
-                rules: [{
-                  instruction: InstructionType.INSERT,
-                  condition: 'name.text="hello"',
-                  index: '$ ~> $map(function($v, $i) { $v.name.text = \'hello\' ? $i })',
-                }],
-              },
-              kind: ts.SyntaxKind.FunctionDeclaration,
-              name: {
-                kind: ts.SyntaxKind.Identifier,
-                text: 'hello',
-              },
-              parameters: [],
-            },
-          ],
-        },
-        sourceFileContents: `
-          function hello() {}
-        `,
-      },
-      {
-        name: createTestName(
-          'should not define an INSERT instruction if',
-          'the field is an array of nodes',
-          'the rules does not evaluate to an instruction',
-        ),
-        input: {
-          kind: ts.SyntaxKind.SourceFile,
-          statements: [
-            {
-              __instructions: {
-                rules: [{
-                  instruction: InstructionType.INSERT,
-                  condition: 'name.text="does-not-exist"',
-                  index: '$ ~> $map(function($v, $i) { $v.name.text = \'hello\' ? $i })',
-                }],
-              },
-              kind: ts.SyntaxKind.FunctionDeclaration,
-              name: {
-                kind: ts.SyntaxKind.Identifier,
-                text: 'hello',
-              },
-              parameters: [],
-            },
-          ],
-        },
-        sourceFileContents: `
-          function hello() {}
-        `,
-      },
-      {
-        name: createTestName(
-          'should throw an error if',
-          'the field is an array of nodes',
-          'a rule evaluates to an INSERT instruction',
-          'a string index is evaluated to a non-integer',
-        ),
-        input: {
-          kind: ts.SyntaxKind.SourceFile,
-          statements: [
-            {
-              __instructions: {
-                rules: [{
-                  instruction: InstructionType.INSERT,
-                  condition: 'name.text="hello"',
-                  index: 'name.text',
-                }],
-              },
-              kind: ts.SyntaxKind.FunctionDeclaration,
-              name: {
-                kind: ts.SyntaxKind.Identifier,
-                text: 'hello',
-              },
-              parameters: [],
-            },
-          ],
-        },
-        error: {
-          prototype: TypeError,
-          message: 'Invalid result, must be integer within the array length: NaN',
-        },
-        sourceFileContents: `
-          function hello() {}
-        `,
-      },
-      {
-        name: createTestName(
-          'should throw an error if',
-          'the field is an array of nodes',
-          'a rule evaluates to an INSERT instruction',
-          'a string index is evaluated to an integer greater than the length of the array of nodes',
-        ),
-        input: {
-          kind: ts.SyntaxKind.SourceFile,
-          statements: [
-            {
-              __instructions: {
-                rules: [{
-                  instruction: InstructionType.INSERT,
-                  condition: 'name.text="hello"',
-                  index:
-                    '$ ~> $map(function($v, $i) { $v.name.text = \'hello\' ? $i + 6 })',
-                }],
-              },
-              kind: ts.SyntaxKind.FunctionDeclaration,
-              name: {
-                kind: ts.SyntaxKind.Identifier,
-                text: 'hello',
-              },
-              parameters: [],
-            },
-          ],
-        },
-        error: {
-          prototype: TypeError,
-          message: 'Invalid result, must be integer within the array length: 6',
-        },
-        sourceFileContents: `
-          function hello() {}
-        `,
-      },
-    ];
-
-    for (const definition of definitions) {
-      blocks.it(definition.name, async (t) => {
-        // Arrange
-        const project = new tsm.Project();
-        const sourceFile = project.createSourceFile(
-          `${crypto.randomUUID()}.ts`,
-          definition.sourceFileContents,
-        );
-
-        if (definition.error) {
-          // Act
-          const result = assertThrows(() =>
-            generateInstructions(sourceFile, definition.input)
-          );
-
-          // Assert
-          assertIsError(result, definition.error.prototype, definition.error.message);
-        } else {
-          // Act
-          const result = generateInstructions(sourceFile, definition.input);
-
-          // Assert
-          await assertSnapshot(t, sanitiseInstructions(result));
-        }
-      });
-    }
   });
 });
